Refactoring to make this more manageable
