+++ Cargo.toml
@@ -33,11 +33,11 @@ bytes = "1.3.0"
 
 [dev-dependencies]
 rust-spice = "0.7.4"
-parquet = "30.0.0"
-arrow = "30.0.0"
+parquet = "30.0.1"
+arrow = "30.0.1"
 criterion = "0.3"
 iai = "0.1"
-polars = {version = "0.26", features = ["lazy", "parquet"]}
+polars = {version = "0.26.1", features = ["lazy", "parquet"]}
 rayon = "1.6.1"
 
 [features]
