+++ data/variable-seg-size-hermite.bsp
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1ce8eaf68c8886b89ebd2a9b13f8a21e68ab72a6c2de5cad2676a152ef275b66
-size 82944
+oid sha256:92d8789c09ddff7f774715fe1f5ac2284c3ebd92120c3554f7945ba06b799378
+size 84992
