+++ src/naif/daf/mod.rs
@@ -15,13 +15,17 @@ use zerocopy::{AsBytes, FromBytes};
 pub(crate) const RCRD_LEN: usize = 1024;
 #[allow(clippy::module_inception)]
 pub mod daf;
-pub mod recordtypes;
+pub mod file_record;
+pub mod name_record;
+pub mod summary_record;
 
 pub use daf::DAF;
 
 use crate::prelude::AniseError;
 use core::fmt::Debug;
-pub use recordtypes::{DAFFileRecord, DAFSummaryRecord, NameRecord};
+pub use file_record::FileRecord;
+pub use name_record::NameRecord;
+pub use summary_record::SummaryRecord;
 
 pub trait NAIFRecord: AsBytes + FromBytes + Sized + Default + Debug {
     const SIZE: usize = core::mem::size_of::<Self>();
@@ -30,8 +34,14 @@ pub trait NAIFRecord: AsBytes + FromBytes + Sized + Default + Debug {
 pub trait NAIFSummaryRecord: NAIFRecord + Copy {
     fn start_index(&self) -> usize;
     fn end_index(&self) -> usize;
+    /// Returns the start epoch in high precision Epoch
     fn start_epoch(&self) -> Epoch;
+    /// Returns the end epoch in high precision Epoch
     fn end_epoch(&self) -> Epoch;
+    /// Returns the start epoch in TDB seconds
+    fn start_epoch_et_s(&self) -> f64;
+    /// Returns the end epoch in TDB seconds
+    fn end_epoch_et_s(&self) -> f64;
     /// Returns whatever is the ID of this summary record.
     fn id(&self) -> i32;
     fn is_empty(&self) -> bool {
@@ -43,6 +53,9 @@ pub trait NAIFDataSet<'a>: Sized + Display {
     /// The underlying record representation
     type RecordKind: NAIFDataRecord<'a>;
 
+    /// The summary record supported by this data set
+    type SummaryKind: NAIFSummaryRecord;
+
     /// The state that is returned from an evaluation of this data set
     type StateKind;
 
@@ -51,7 +64,11 @@ pub trait NAIFDataSet<'a>: Sized + Display {
 
     fn nth_record(&self, n: usize) -> Result<Self::RecordKind, AniseError>;
 
-    fn evaluate(&self, epoch: Epoch, start_epoch: Epoch) -> Result<Self::StateKind, AniseError>;
+    fn evaluate(
+        &self,
+        epoch: Epoch,
+        summary: &Self::SummaryKind,
+    ) -> Result<Self::StateKind, AniseError>;
 
     /// Checks the integrity of this data set, returns an error if the data has issues.
     fn check_integrity(&self) -> Result<(), AniseError>;
