+++ src/naif/pck/mod.rs
@@ -47,4 +47,12 @@ impl NAIFSummaryRecord for BPCSummaryRecord {
     fn id(&self) -> i32 {
         self.frame_id
     }
+
+    fn start_epoch_et_s(&self) -> f64 {
+        self.start_epoch_et_s
+    }
+
+    fn end_epoch_et_s(&self) -> f64 {
+        self.end_epoch_et_s
+    }
 }
