+++ tests/ephemerides/validation/type13_hermite.rs
@@ -10,6 +10,7 @@
 
 use super::{compare::*, validate::Validation};
 
+#[ignore = "Requires Rust SPICE -- must be executed serially"]
 #[test]
 fn validate_hermite_type13_from_gmat() {
     let file_name = "spk-type13-validation-even-seg-size".to_string();
@@ -31,8 +32,10 @@ fn validate_hermite_type13_from_gmat() {
     validator.validate();
 }
 
+#[ignore = "Requires Rust SPICE -- must be executed serially"]
 #[test]
 fn validate_hermite_type13_with_varying_segment_sizes() {
+    // ISSUE: This file is corrupt. I messed up the rewrite.
     let file_name = "spk-type13-validation-variable-seg-size".to_string();
     let comparator = CompareEphem::new(
         vec!["data/variable-seg-size-hermite.bsp".to_string()],
